+++ webapp/store/survey/categories/index.js
@@ -1,0 +1,5 @@
+import * as CategoriesState from './state'
+import * as CategoriesActions from './actions'
+import CategoryReducer from './reducer'
+
+export { CategoriesActions, CategoriesState, CategoryReducer }
