+++ webapp/store/survey/categories/state.js
@@ -9,14 +9,14 @@ import * as ObjectUtils from '@core/objectUtils'
 
 // Category
 
-export const assocCategory = category => R.assoc(Category.getUuid(category), category)
+export const assocCategory = (category) => R.assoc(Category.getUuid(category), category)
 
-export const dissocCategory = category => R.dissoc(Category.getUuid(category))
+export const dissocCategory = (category) => R.dissoc(Category.getUuid(category))
 
 export const assocCategoryProp = (category, key, value) =>
   R.pipe(
     R.assocPath([Category.getUuid(category), ObjectUtils.keys.props, key], value),
-    R.dissocPath([Category.getUuid(category), Validation.keys.validation, Validation.keys.fields, key]),
+    R.dissocPath([Category.getUuid(category), Validation.keys.validation, Validation.keys.fields, key])
   )
 
 // Category level
@@ -31,7 +31,7 @@ export const assocCategoryLevelProp = (category, level, key, value) =>
         ObjectUtils.keys.props,
         key,
       ],
-      value,
+      value
     ),
     R.dissocPath([
       Category.getUuid(category),
@@ -40,7 +40,7 @@ export const assocCategoryLevelProp = (category, level, key, value) =>
       Validation.keys.validation,
       Validation.keys.fields,
       key,
-    ]),
+    ])
   )
 
 export const dissocCategoryLevel = (category, level) =>
