+++ webapp/store/survey/index.js
@@ -1,0 +1,7 @@
+import * as SurveyActions from './actions'
+import * as SurveyState from './surveyState'
+import SurveyReducer from './reducer'
+
+export { SurveyActions, SurveyState, SurveyReducer }
+
+export { CategoriesState, CategoriesActions, CategoryReducer } from './categories'
