+++ webapp/store/survey/surveyInfo/reducer.js
