+++ webapp/store/survey/taxonomies/actions.js
