+++ webapp/store/survey/taxonomies/reducer.js
